refactor: BlankTest, PositionTest 패키지 이동
piece 패키지
