+++ src/test/java/chess/domain/piece/BlankTest.java
@@ -1,12 +1,9 @@
-package chess.domain;
+package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import chess.domain.piece.Blank;
-import chess.domain.piece.Position;
 
 public class BlankTest {
 	@Test
