+++ src/test/java/chess/domain/piece/PositionTest.java
@@ -1,11 +1,9 @@
-package chess.domain;
+package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import chess.domain.piece.Position;
 
 public class PositionTest {
 	@Test
